BlobHeaderV1::finalize()
